[c++] Move installed header files one level deeper
