+++ apis/python/src/tiledbsoma/query_condition.cc
@@ -41,8 +41,8 @@
 
 // #include "util.h"
 #include <tiledb/tiledb> // C++
+#include <tiledbsoma/tiledbsoma>
 
-#include "tiledbsoma/logger_public.h"
 #define TPY_ERROR_LOC(m) throw tiledbsoma::TileDBSOMAError(m);
 
 #if TILEDB_VERSION_MAJOR == 2 && TILEDB_VERSION_MINOR >= 2
