+++ libtiledbsoma/src/soma/logger_public.h
