+++ libtiledbsoma/src/soma/managed_query.h
@@ -38,9 +38,9 @@
 
 #include <tiledb/tiledb>
 
-#include "utils/array_buffers.h"
-#include "utils/column_buffer.h"
-#include "utils/common.h"
+#include "../utils/array_buffers.h"
+#include "../utils/column_buffer.h"
+#include "../utils/common.h"
 
 namespace tiledbsoma {
