+++ libtiledbsoma/src/tiledbsoma/tiledbsoma
@@ -42,14 +42,14 @@
 // #include "soma_object.h"
 // #include "soma_sparse_ndarray.h"
 
-#include <utils/arrow_adapter.h>
-#include <tiledbsoma/logger_public.h>
-#include <utils/array_buffers.h>
-#include <utils/column_buffer.h>
-#include <utils/common.h>
-#include <utils/stats.h>
-#include <utils/version.h>
-#include <soma/managed_query.h>
-#include <soma/soma_array.h>
+#include "utils/arrow_adapter.h"
+#include "utils/array_buffers.h"
+#include "utils/column_buffer.h"
+#include "utils/common.h"
+#include "utils/stats.h"
+#include "utils/version.h"
+#include "soma/logger_public.h"
+#include "soma/managed_query.h"
+#include "soma/soma_array.h"
 
 #endif
